Add recommendation to sort tokens
The container is specified to be CBOR encoded, but the tokens within the container are themselves an _array_ of bytes.

It means that you can add the _same_ tokens to the container in a _different_ order and get a _different encoded representation_.

This PR updates the wording to recommend that the tokens are ordered bytewise, which, when followed will ensure that there is only 1 set of bytes that correspond to the container contents. i.e. adding determinism.

Signed-off-by: ash <alan138@gmail.com>